Split track-related classes out of HitStructures.h/cc into a new TrackStructures.h/cc file-set.
