+++ RecoTracker/MkFitCore/src/CandCloner.cc
@@ -1,6 +1,5 @@
 #include "CandCloner.h"
 
-#include "RecoTracker/MkFitCore/interface/HitStructures.h"
 #include "RecoTracker/MkFitCore/interface/IterationConfig.h"
 
 //#define DEBUG
